+++ docs/changelog.rst
@@ -6,6 +6,12 @@ Changelog
 `Unreleased`_
 -------------
 
+Changed
+~~~~~~~
+
+- Schemas with GET endpoints accepting body are allowed now if schema validation is disabled (via ``--validate-schema=false`` for example).
+  The usecase is for tools like ElasticSearch that use GET requests with non empty bodies. `#383`_
+
 `0.23.7`_ - 2020-01-30
 ----------------------
 
@@ -656,6 +662,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#383: https://github.com/kiwicom/schemathesis/issues/383
 .. _#381: https://github.com/kiwicom/schemathesis/issues/381
 .. _#378: https://github.com/kiwicom/schemathesis/issues/378
 .. _#376: https://github.com/kiwicom/schemathesis/issues/376